converted mediaId.categoryId from `Int` to ´String´, to be compatible with Spotify string ids

+++ core/src/main/java/dev/olog/core/interactor/DeleteUseCase.kt
@@ -28,8 +28,8 @@ class DeleteUseCase @Inject constructor(
 
     private suspend fun handleCategory(mediaId: Category) {
         return when (mediaId.category) {
-            PODCASTS_PLAYLIST -> podcastPlaylistGateway.deletePlaylist(mediaId.categoryId)
-            PLAYLISTS -> playlistGateway.deletePlaylist(mediaId.categoryId)
+            PODCASTS_PLAYLIST -> podcastPlaylistGateway.deletePlaylist(mediaId.categoryId.toLong())
+            PLAYLISTS -> playlistGateway.deletePlaylist(mediaId.categoryId.toLong())
             else -> {
                 val songList = getSongListByParamUseCase(mediaId)
                 trackGateway.deleteGroup(songList.map { it.id })
