+++ core/src/main/java/dev/olog/core/interactor/ObserveRelatedArtistsUseCase.kt
@@ -22,9 +22,9 @@ class ObserveRelatedArtistsUseCase @Inject constructor(
     operator fun invoke(mediaId: MediaId): Flow<List<Artist>> {
         return when (mediaId.category) {
             MediaIdCategory.FOLDERS -> folderGateway.observeRelatedArtists(mediaId.categoryId)
-            MediaIdCategory.PLAYLISTS -> playlistGateway.observeRelatedArtists(mediaId.categoryId)
-            MediaIdCategory.GENRES -> genreGateway.observeRelatedArtists(mediaId.categoryId)
-            MediaIdCategory.PODCASTS_PLAYLIST -> podcastPlaylistGateway.observeRelatedArtists(mediaId.categoryId)
+            MediaIdCategory.PLAYLISTS -> playlistGateway.observeRelatedArtists(mediaId.categoryId.toLong())
+            MediaIdCategory.GENRES -> genreGateway.observeRelatedArtists(mediaId.categoryId.toLong())
+            MediaIdCategory.PODCASTS_PLAYLIST -> podcastPlaylistGateway.observeRelatedArtists(mediaId.categoryId.toLong())
             else -> flowOf(listOf())
         }
     }
